runtime_demo 接入 + debug_log 预算修复 + VFS/term 安全加固
- 接入 runtime_demo 服务：Makefile 4 处（USER_C/ENTRY_OBJ/SVC_NAMES/LINK_RULE）、
  kernel blob 注册、shell spawn 支持 blob 名参数（默认 hello 向后兼容）
- 修复 sys_debug_log token bucket：惰性 refill 致 init 回归尾部后预算冻结，
  单 tick 突发 ~1.7KB 的 runtime_demo 输出被截断吞没。现按 tick refill 2048
  并允许 4096 突发（原 512 硬重置），持续洪泛仍按 tick 限速
- signal.h 重写为兼容头（规范定义移至 libos/syscalls.h），signal() 返回 SIG_ERR
  而非裸负数；libos map_memory/unmap_memory 改 uint64_t 参数（ASLR 堆区可超 INT_MAX）
- VFS 安全加固：mount/unmount 绑定挂载者 subject（防任意卸载 DoS）、close/revoke
  校验调用者身份、bookmark resolve 用 live_id 不篡改锚定记录、extent 拷贝失败回滚
  释放新块、新增 VFS_OP_LIST_VOLUMES 卷枚举（"/" 根视图）
- term Powerbox 面板恢复改为仅恢复面板矩形（保留面板期间 shell 输出与提示符）
- init 新增 KBD 焦点所有权往返回归测试（TAKE/RELEASE/非 owner ERR_NOCAP）
- 清理 build/ 与 .omo/ 历史误跟踪产物（.gitignore 已声明，仅取消索引跟踪）

+++ user/runtime/malloc.c
@@ -156,7 +156,7 @@ static int heap_grow(size_t need) {
 
     /* map_memory returns the virtual address on success, 0 on failure.
      * The second argument (offset) is the desired virtual address. */
-    void *addr = map_memory(cap, (int)s_next_virt, (int)chunk, PROT_READ | PROT_WRITE);
+    void *addr = map_memory(cap, s_next_virt, chunk, PROT_READ | PROT_WRITE);
     if (!addr) {
         cap_revoke(cap);
         return -1;
